% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/contains_entities.R
\name{contains_entities}
\alias{contains_entities}
\title{Contains Entities}
\usage{
contains_entities(input, entities, model.list)
}
\arguments{
\item{input}{A character vector of text to be analysed.}

\item{entities}{A character vector of named entities.}

\item{model.list}{A model list, produced using load_nlp_models.}
}
\value{
A list containing: (a) a logical vector specifying whether each entity named in 'entities' was found in the text, and (b) a list of the identified entities.
}
\description{
Determines whether a text input contains any entities of a pre-specified type, and reports those entities.
}
\examples{
a <- load_nlp_models(entities=c("sentence","word","POS","person"))
input <- "This is a test."
entities <- c("POS/NNP","person")
contains_entities(input,entities,model.list=a)
}

