% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/entities_in_common.R
\name{entities_in_common}
\alias{entities_in_common}
\title{Entities in Common}
\usage{
entities_in_common(input1, input2, model.list)
}
\arguments{
\item{input1}{The first character vector of text to be compared.}

\item{input2}{The second character vector of text to be compared.}

\item{model.list}{A model list, produced using load_nlp_models.}
}
\description{
Finds a set of common named entities among two text inputs.
}

