% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract_entities.R
\name{extract_entities}
\alias{extract_entities}
\title{Extract Shared Entities}
\usage{
extract_entities(input, model.list)
}
\arguments{
\item{input}{A character vector of text to be analysed.}

\item{model.list}{A model list, produced using load_nlp_models.}
}
\description{
Extracts a set of named entities from a text input according to a set of NLP models.
}

