% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_nlp_models.R
\name{load_nlp_models}
\alias{load_nlp_models}
\title{Load NLP Models}
\usage{
load_nlp_models(entities)
}
\arguments{
\item{entities}{Character vector of entities.}
}
\value{
A list of openNLP models
}
\description{
Loads a set of OpenNLP models according to an optional character vector of entities.
}
\examples{
a <- load_nlp_models(c("sentence","word","POS","person","organization","location","date","time","percentage"))
}

