% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nlp_engine.R
\name{nlp_engine}
\alias{nlp_engine}
\title{NLP Engine}
\usage{
nlp_engine(input, model.list)
}
\arguments{
\item{input}{A character vector of text to be analysed.}

\item{model.list}{A model list, produced using load_nlp_models.}
}
\description{
Builds an Annotated Plaintext Document from a character vector input.
}

