% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shared_entities.R
\name{shared_entities}
\alias{shared_entities}
\title{Identify Shared Named Entities}
\usage{
shared_entities(input1, input2, model.list)
}
\arguments{
\item{input1}{The first character vector of text to be compared.}

\item{input2}{The second character vector of text to be compared.}

\item{model.list}{A model list, produced using load_nlp_models.}
}
\description{
Finds a set of named entities that are shared by two text inputs.
}

